+++ index.js
@@ -1,7 +1,7 @@
-'use strict';
+"use strict";
 
-const { path: projectRootPath } = require('app-root-path');
-const readPkgUp = require('read-pkg-up');
+const { path: projectRootPath } = require("app-root-path");
+const readPkgUp = require("read-pkg-up");
 
 const {
   packageJson: {
@@ -15,11 +15,11 @@ const {
 
 const env = {
   browser: !!browserslist,
-  babel: !!devDependencies['@babel/core'] || !!dependencies.next,
+  babel: !!devDependencies["@babel/core"] || !!dependencies.next,
   prettier: !!devDependencies.prettier,
   react: !!peerDependencies.react || !!dependencies.react,
   next: !!dependencies.next,
-  jsdocMd: !!devDependencies['jsdoc-md'],
+  jsdocMd: !!devDependencies["jsdoc-md"],
 };
 
 /**
@@ -38,10 +38,10 @@ checkDevDependencies([
   // Although the fact this config is being used implies ESLint is present, it
   // may still be missing from dev dependencies in the case of a global
   // installation or an editor plugin.
-  'eslint',
-  'eslint-plugin-node',
-  'eslint-plugin-import',
-  'eslint-plugin-jsdoc',
+  "eslint",
+  "eslint-plugin-node",
+  "eslint-plugin-import",
+  "eslint-plugin-jsdoc",
 ]);
 
 /**
@@ -49,36 +49,36 @@ checkDevDependencies([
  * @see [jsdoc-md docs](https://github.com/jaydenseric/jsdoc-md#tag-subset).
  */
 const JSDOC_MD_SUPPORTED_TAGS = [
-  'desc',
-  'description',
-  'kind',
-  'name',
-  'typedef',
-  'callback',
-  'type',
-  'prop',
-  'property',
-  'arg',
-  'argument',
-  'param',
-  'return',
-  'returns',
-  'emits',
-  'fires',
-  'see',
-  'example',
-  'ignore',
+  "desc",
+  "description",
+  "kind",
+  "name",
+  "typedef",
+  "callback",
+  "type",
+  "prop",
+  "property",
+  "arg",
+  "argument",
+  "param",
+  "return",
+  "returns",
+  "emits",
+  "fires",
+  "see",
+  "example",
+  "ignore",
 ];
 
 /**
  * Preferred JSDoc tag names.
  */
 const JSDOC_TAG_NAME_PREFERENCE = {
-  property: 'prop',
-  arg: 'param',
-  argument: 'param',
-  return: 'returns',
-  emits: 'fires',
+  property: "prop",
+  arg: "param",
+  argument: "param",
+  return: "returns",
+  emits: "fires",
 };
 
 /**
@@ -87,7 +87,7 @@ const JSDOC_TAG_NAME_PREFERENCE = {
  * @returns {object} Tag name preference.
  */
 function jsdocMdTagNamePreference() {
-  const { jsdocTags } = require('eslint-plugin-jsdoc/dist/tagNames');
+  const { jsdocTags } = require("eslint-plugin-jsdoc/dist/tagNames");
   const tagNamePreference = {
     ...JSDOC_TAG_NAME_PREFERENCE,
   };
@@ -112,8 +112,8 @@ const config = {
         ? jsdocMdTagNamePreference()
         : JSDOC_TAG_NAME_PREFERENCE,
       preferredTypes: {
-        '.<>': '<>',
-        '[]': 'Array<>',
+        ".<>": "<>",
+        "[]": "Array<>",
       },
     },
   },
@@ -121,70 +121,70 @@ const config = {
     es2021: true,
     node: true,
   },
-  plugins: ['jsdoc'],
+  plugins: ["jsdoc"],
   extends: [
-    'eslint:recommended',
-    'plugin:node/recommended',
-    'plugin:import/recommended',
+    "eslint:recommended",
+    "plugin:node/recommended",
+    "plugin:import/recommended",
   ],
   rules: {
-    'no-return-await': 'error',
-    'no-console': [
-      'error',
+    "no-return-await": "error",
+    "no-console": [
+      "error",
       {
         allow: [
-          'error',
-          'group',
-          'groupCollapsed',
-          'groupEnd',
-          'info',
-          'table',
-          'warn',
+          "error",
+          "group",
+          "groupCollapsed",
+          "groupEnd",
+          "info",
+          "table",
+          "warn",
         ],
       },
     ],
-    curly: ['error', 'multi'],
-    'sort-imports': ['error', { ignoreDeclarationSort: true }],
-    'require-unicode-regexp': 'error',
-    strict: 'error',
+    curly: ["error", "multi"],
+    "sort-imports": ["error", { ignoreDeclarationSort: true }],
+    "require-unicode-regexp": "error",
+    strict: "error",
 
-    'node/file-extension-in-import': ['error', 'always'],
+    "node/file-extension-in-import": ["error", "always"],
 
-    'import/first': 'error',
-    'import/newline-after-import': 'error',
-    'import/no-useless-path-segments': 'error',
-    'import/no-unresolved': 'off',
-    'import/order': [
-      'error',
+    "import/first": "error",
+    "import/newline-after-import": "error",
+    "import/no-useless-path-segments": "error",
+    "import/no-unresolved": "off",
+    "import/order": [
+      "error",
       {
-        alphabetize: { order: 'asc' },
-        'newlines-between': 'never',
+        alphabetize: { order: "asc" },
+        "newlines-between": "never",
       },
     ],
 
-    'jsdoc/check-alignment': 'error',
-    'jsdoc/check-param-names': 'error',
-    'jsdoc/check-property-names': 'error',
-    'jsdoc/check-tag-names': 'error',
-    'jsdoc/check-types': 'error',
-    'jsdoc/empty-tags': 'error',
-    'jsdoc/implements-on-classes': 'error',
-    'jsdoc/newline-after-description': ['error', 'never'],
-    'jsdoc/require-description': 'error',
-    'jsdoc/require-jsdoc': 'error',
-    'jsdoc/require-param': 'error',
-    'jsdoc/require-param-description': 'error',
-    'jsdoc/require-param-name': 'error',
-    'jsdoc/require-param-type': 'error',
-    'jsdoc/require-property': 'error',
-    'jsdoc/require-property-description': 'error',
-    'jsdoc/require-property-name': 'error',
-    'jsdoc/require-property-type': 'error',
-    'jsdoc/require-returns': 'error',
-    'jsdoc/require-returns-check': 'error',
-    'jsdoc/require-returns-description': 'error',
-    'jsdoc/require-returns-type': 'error',
-    'jsdoc/valid-types': 'error',
+    "jsdoc/check-alignment": "error",
+    "jsdoc/check-param-names": "error",
+    "jsdoc/check-property-names": "error",
+    "jsdoc/check-tag-names": "error",
+    "jsdoc/check-types": "error",
+    "jsdoc/empty-tags": "error",
+    "jsdoc/implements-on-classes": "error",
+    "jsdoc/newline-after-description": ["error", "never"],
+    "jsdoc/require-description": "error",
+    "jsdoc/require-jsdoc": "error",
+    "jsdoc/require-param": "error",
+    "jsdoc/require-param-description": "error",
+    "jsdoc/require-param-name": "error",
+    "jsdoc/require-param-type": "error",
+    "jsdoc/require-property": "error",
+    "jsdoc/require-property-description": "error",
+    "jsdoc/require-property-name": "error",
+    "jsdoc/require-property-type": "error",
+    "jsdoc/require-returns": "error",
+    "jsdoc/require-returns-check": "error",
+    "jsdoc/require-returns-description": "error",
+    "jsdoc/require-returns-type": "error",
+    "jsdoc/valid-types": "error",
   },
 
   // These base options apply to all linted files, including .js and .jsx.
@@ -199,13 +199,13 @@ const config = {
     // try to suit the project environment. See:
     // https://nodejs.org/api/packages.html#packages_type
     sourceType: type
-      ? type === 'module'
-        ? 'module'
-        : 'script'
+      ? type === "module"
+        ? "module"
+        : "script"
       : // Next.js projects allow ESM in non .mjs files.
       env.next
-      ? 'module'
-      : 'script',
+      ? "module"
+      : "script",
   },
 
   // Enforce file extension specific Node.js standards. eslint-plugin-node
@@ -215,15 +215,15 @@ const config = {
   // unless overridden by the consumer’s config.
   overrides: [
     {
-      files: ['*.cjs'],
+      files: ["*.cjs"],
       parserOptions: {
-        sourceType: 'script',
+        sourceType: "script",
       },
     },
     {
-      files: ['*.mjs'],
+      files: ["*.mjs"],
       parserOptions: {
-        sourceType: 'module',
+        sourceType: "module",
       },
     },
   ],
@@ -233,25 +233,25 @@ if (env.browser) {
   config.env.browser = true;
 
   if (!env.babel) {
-    checkDevDependencies(['eslint-plugin-compat']);
-    config.extends.push('plugin:compat/recommended');
+    checkDevDependencies(["eslint-plugin-compat"]);
+    config.extends.push("plugin:compat/recommended");
   }
 }
 
 if (env.babel || !env.browser) {
   // It would be nice to also prefer modern ES syntax for browser projects, when
   // available in all browsers supported in the project’s browserslist config.
-  config.rules['prefer-destructuring'] = 'error';
-  config.rules['object-shorthand'] = [
-    'error',
-    'always',
+  config.rules["prefer-destructuring"] = "error";
+  config.rules["object-shorthand"] = [
+    "error",
+    "always",
     { avoidExplicitReturnArrows: true },
   ];
 }
 
 if (env.babel) {
-  checkDevDependencies(['@babel/eslint-parser']);
-  config.parser = '@babel/eslint-parser';
+  checkDevDependencies(["@babel/eslint-parser"]);
+  config.parser = "@babel/eslint-parser";
 
   // Prevent a parsing error when linting a file not under the scope of a Babel
   // config file.
@@ -260,53 +260,53 @@ if (env.babel) {
   // Assume all unsupported Node.js features used are transpiled. It would be
   // nice if there was a way to check Babel config and only disable checking
   // features known to be transpiled.
-  config.rules['node/no-unsupported-features/es-builtins'] = 'off';
-  config.rules['node/no-unsupported-features/es-syntax'] = 'off';
+  config.rules["node/no-unsupported-features/es-builtins"] = "off";
+  config.rules["node/no-unsupported-features/es-syntax"] = "off";
 }
 
 if (env.react) {
-  checkDevDependencies(['eslint-plugin-react', 'eslint-plugin-react-hooks']);
-  config.extends.push('plugin:react/recommended');
+  checkDevDependencies(["eslint-plugin-react", "eslint-plugin-react-hooks"]);
+  config.extends.push("plugin:react/recommended");
 
   // Prevents an eslint-plugin-react warning, see:
   // https://github.com/yannickcr/eslint-plugin-react/issues/1955#issuecomment-450771510
-  config.settings.react = { version: 'detect' };
+  config.settings.react = { version: "detect" };
 
-  config.rules['react/prop-types'] = 'off';
-  config.rules['react/jsx-boolean-value'] = 'error';
-  config.rules['react/jsx-curly-brace-presence'] = 'error';
-  config.rules['react/jsx-fragments'] = 'error';
-  config.rules['react/jsx-no-useless-fragment'] = 'error';
-  config.rules['react/no-array-index-key'] = 'error';
+  config.rules["react/prop-types"] = "off";
+  config.rules["react/jsx-boolean-value"] = "error";
+  config.rules["react/jsx-curly-brace-presence"] = "error";
+  config.rules["react/jsx-fragments"] = "error";
+  config.rules["react/jsx-no-useless-fragment"] = "error";
+  config.rules["react/no-array-index-key"] = "error";
 
-  config.plugins.push('react-hooks');
-  config.rules['react-hooks/rules-of-hooks'] = 'error';
-  config.rules['react-hooks/exhaustive-deps'] = 'error';
+  config.plugins.push("react-hooks");
+  config.rules["react-hooks/rules-of-hooks"] = "error";
+  config.rules["react-hooks/exhaustive-deps"] = "error";
 }
 
 if (env.prettier) {
-  checkDevDependencies(['eslint-config-prettier', 'eslint-plugin-prettier']);
-  config.extends.push('plugin:prettier/recommended');
+  checkDevDependencies(["eslint-config-prettier", "eslint-plugin-prettier"]);
+  config.extends.push("plugin:prettier/recommended");
 }
 
 if (env.next) {
-  checkDevDependencies(['@next/eslint-plugin-next']);
-  config.extends.push('plugin:@next/next/recommended');
+  checkDevDependencies(["@next/eslint-plugin-next"]);
+  config.extends.push("plugin:@next/next/recommended");
 
   // Don’t warn when vanilla `img` elements are used in JSX as they have many
   // valid uses.
-  config.rules['@next/next/no-img-element'] = 'off';
+  config.rules["@next/next/no-img-element"] = "off";
 
   // Next.js projects allow ESM in non .mjs files.
-  config.rules['node/no-extraneous-import'] = 'error';
-  config.rules['node/no-missing-import'] = 'error';
-  config.rules['node/no-unpublished-import'] = 'error';
+  config.rules["node/no-extraneous-import"] = "error";
+  config.rules["node/no-missing-import"] = "error";
+  config.rules["node/no-unpublished-import"] = "error";
 
   // Next.js projects allow extensionless import specifiers.
-  config.rules['node/file-extension-in-import'] = 'off';
+  config.rules["node/file-extension-in-import"] = "off";
 
   // Next.js uses https://npm.im/babel-plugin-react-require.
-  config.rules['react/react-in-jsx-scope'] = 'off';
+  config.rules["react/react-in-jsx-scope"] = "off";
 }
 
 module.exports = config;